requirements.txt
fastapi~=0.123.5
fastapi[standard]~=0.123.5
requests~=2.32.5
python-dotenv~=1.2.1
uvicorn~=0.38.0